refactor: code quality pass — fix 5 correctness/perf issues
## Fixes

### ui/app.py
- Remove unused `get_prediction_label` import (imported but never called)
- Replace `assert` in `_get_singletons()` with explicit `RuntimeError`
  (assert is silently disabled when Python runs with -O flag)
- Fix `load_dashboard()` called 4× during `build_app()` startup:
  compute once, unpack to local vars — avoids redundant sklearn + DB work

### detector/model.py
- `_features_to_matrix`: float32 → float64
  (sklearn's StandardScaler operates in float64; float32 causes silent
   precision loss in SVM's RBF kernel distance computation)

### requirements.txt + requirements-optional.txt (new)
- Split deps: core (required) vs optional (pydub, opensmile, faster-whisper)
  These were listed as required but the code handles their absence gracefully

### tests/test_detector.py
- Add TestPretrainedModel: 4 tests covering pretrained model lifecycle
- Add TestFeatureMatrix: float64 dtype + NaN-free fill assertions
- Total tests: 20 → 31 (all passing)

### .gitignore
- Exclude __pycache__, .venv, data/*.db, models/*.joblib

+++ detector/model.py
@@ -378,7 +378,7 @@ class LieDetectorModel:
         """
         n = len(features_list)
         m = len(self.feature_names)
-        X = np.zeros((n, m), dtype=np.float32)
+        X = np.zeros((n, m), dtype=np.float64)  # float64: sklearn StandardScaler precision
         for i, fd in enumerate(features_list):
             for j, name in enumerate(self.feature_names):
                 val = fd.get(name, 0.0)
